+++ gallery/gallery.js
@@ -1,7 +1,7 @@
 import { setExpireTimeAfterReloading } from "../utilities/token/setToken.js";
-import { reRenderGalleryPage } from "../utilities/render/render.js";
 import { UrlManipulationService } from "../services/url-manipulation.service.js";
 import { renderGalleryPage } from "/gallery/handler.js";
+import { reRenderGalleryPage } from "/gallery/handler.js"
 
 const urlService = new UrlManipulationService();
 
@@ -14,4 +14,4 @@ if (!urlService.tokenExists()) {
 await renderGalleryPage(); // handler
 
 const pages = document.getElementById('pages');
-pages.onclick = await reRenderGalleryPage();
+pages.onclick = await reRenderGalleryPage(); // handler
